+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 FEATURES
 * modules/mesh-task: Add `envoy_public_listener_port` variable to set Envoy's public listener port.
+* modules/acl-controller: Add `additional_execution_role_policies` variable to support attaching custom policies to the task's execution role.
 
 ## 0.5.1 (July 29, 2022)